made some updates to bugs and also sound playing

+++ transcriber.py
@@ -22,11 +22,18 @@ from pynput.mouse import Button, Controller
 import cv2
 import numpy as np
 from threading import Thread
+from dotenv import load_dotenv
+
+#Common transcription variations of your name. Case sensitive.
+keywordOne = 'edward'
+keywordTwo ='Edward'
+keywordThree='eward'
 
 is_bot_speaking = False
 mouse = Controller()
+load_dotenv()
 ssl._create_default_https_context = ssl._create_unverified_context
-openai.api_key = 'DEFAULT'
+openai.api_key = os.getenv('OPENAI_API_KEY')
 
 def find_image_on_screen(image_file: str):
     screen = np.array(pag.screenshot())
@@ -62,40 +69,45 @@ def trigger_robot(brain_needed, status_queue, is_terminate):
     def run():
         global is_bot_speaking
         is_bot_speaking = True
+        print('is bot speaking set 2 true')
         try:
             print(f'WHAT WANT TO GIVE ROBOT: {brain_needed}')
             status_queue.put("BuellerBot is thinking")
             print("------yepcock---------")
-
-            completion = openai.Completion.create(
-                model="text-davinci-003",
-                prompt=brain_needed,
-                max_tokens=75,
-                temperature=0
+            #the brains of the operation
+            completion = openai.ChatCompletion.create(
+                model="gpt-3.5-turbo",
+                messages=[
+                {"role": "assistant", "content": f'INSTURCTION: You are {keywordOne}, a college student who is very smart and kind of funny. Please provide a brief, professional, and kind of funny response to the following transcript. KEEP IN MIND that the transcript will likely be cut off. TRANSCRIPT: {brain_needed} INSUTRCTION: DO NOT continue where the transcript left off -- instead, form a COMPLETE SENTENCE to respond to the most recent question. RESPONSE:'}
+                ],
+                max_tokens=55,
+                temperature=0.2
             )
-
-            brain_og = completion.choices[0].text
+            #the lobotomy of the operation
+            brain_og = completion.choices[0].message.content
             brain_given = brain_og.replace('\n', '')
             with open('brain_given.txt', 'w') as file:
                 file.write(brain_given.strip())
             with open('transcriptions/transcript.txt', 'a') as file:
-                file.write('\n\nBuellerBot Response For Context: ' + brain_given.strip() + '\n\nTranscript: ')
+                file.write('\n\nPAST RESPONSE: ' + brain_given.strip() + '\n\nTRANSCRIPT:') #allows for follow-up questions. 
             print(f'WHAT ROBOT RESPONSED TO: {brain_needed}')
             print(f'ROBOT ABOUT TO SAY: {brain_given}')
             if is_terminate.value:
                 raise TerminateSignal
             status_queue.put("BuellerBot is unmuting")
             print("ROBOT IS ABOUT TO UNMUTE")
-
+            #search for mute button
             x, y = find_image_on_screen('g_unmute.png')
             mouse.position = (x, y)
             mouse.click(Button.left, 1)
             time.sleep(1)
-            # mouse.click(Button.left, 1)
+            mouse.click(Button.left, 1)
             status_queue.put("BuellerBot is buying time")
             if is_terminate.value:
                 raise TerminateSignal
             t2a("Hey sorry i''m having some audio issues. Is my mic working? Can you hear me?")
+            is_bot_speaking = True
+            print('is bot speaking set 2 true')
             playsound('output.mp3')
             # with open("transcriptions/transcript.txt", "r+") as fx:
             #     fx.truncate()
@@ -107,16 +119,24 @@ def trigger_robot(brain_needed, status_queue, is_terminate):
             t2a(brain_given)
             print("EL API Response Received")
             status_queue.put("BuellerBot is responding")
+            is_bot_speaking = True
+            print('is bot speaking set 2 true')
             playsound('output.mp3')
+            time.sleep(5) #allow it to respond w/o getting caught in the transcript
             status_queue.put("BuellerBot is about to mute")
             print("ROBOT IS ABOUT TO MUTE")
+            is_bot_speaking = True
+            print('is bot speaking set 2 true')
             x, y = find_image_on_screen('g_mute.png')
             mouse.position = (x, y)
             mouse.click(Button.left, 1)
             status_queue.put("BuellerBot is on standby")
             is_bot_speaking = False
+            print('is bot speaking set 2 false')
         finally:
             is_bot_speaking = False
+            print('is bot speaking set 2 false')
+            # Add some delay here to ensure the bot is fully finished before resuming transcription
     Thread(target=run).start()
 
 class TerminateSignal(Exception):
@@ -133,7 +153,6 @@ def start_transcription(status_queue, is_terminate):
         if is_terminate.value:
             print("terminate hit, break function")
             break
-
         # get most recent wav recording in the recordings directory
         try:
             files = sorted(glob.iglob(recordings_dir), key=os.path.getctime, reverse=True)
@@ -141,21 +160,25 @@ def start_transcription(status_queue, is_terminate):
             files = []
         if len(files) < 1:
             continue
-
         latest_recording = files[0]
         latest_recording_filename = latest_recording.split('/')[1]
-
+        if is_bot_speaking:
+                    continue
         if os.path.exists(latest_recording) and not latest_recording in transcribed:
             audio = whisper.load_audio(latest_recording)
             audio = whisper.pad_or_trim(audio)
             mel = whisper.log_mel_spectrogram(audio).to(model.device)
             options = whisper.DecodingOptions(language= 'en', fp16=False)
-
             result = whisper.decode(model, mel, options)
-
+            if is_bot_speaking:
+                continue
             if result.no_speech_prob < 0.5:
+                if is_bot_speaking:
+                    continue
                 print("Live Transcript: " + result.text)
                 with open(config.TRANSCRIPT_FILE, 'a+') as f:
+                    if is_bot_speaking:
+                        continue
                     f.write(result.text)
                     f.truncate()
                     f.seek(0)  
@@ -168,12 +191,16 @@ def start_transcription(status_queue, is_terminate):
 
                 with open("transcriptions/transcript.txt", "r") as fx:
                     brain_needed = fx.read()
-                    if 'edward' in brain_needed or 'eward' in brain_needed or 'Edward' in brain_needed:
+                    if is_bot_speaking:
+                        continue
+                    if keywordOne in brain_needed or keywordTwo in brain_needed or keywordThree in brain_needed:
+                        print('trigger word noted, waiting 1 second for additional context')
+                        time.sleep(1.5) #keep recording for 1.5 seconds in case of additional context
                         trigger_robot(brain_needed, status_queue, is_terminate)
-                        brain_needed = brain_needed.replace('edward', '')
-                        brain_needed = brain_needed.replace('Edward', '')
-                        brain_needed = brain_needed.replace('eward', '')
+                        brain_needed = brain_needed.replace(keywordOne, '')
+                        brain_needed = brain_needed.replace(keywordTwo, '')
+                        brain_needed = brain_needed.replace(keywordThree, '')
                         with open("transcriptions/transcript.txt", "w") as fw:
                             fw.write(brain_needed)
-
+            
             transcribed.append(latest_recording)
